Refresh avgRate on delete User, prevent User from deleting himself

+++ src/controllers/AdminUserController.php
@@ -2,15 +2,18 @@
 
 require_once 'AppController.php';
 require_once  __DIR__.'/../repositories/UserRepository.php';
+require_once  __DIR__.'/../repositories/FilmRepository.php';
 
 class AdminUserController extends AppController {
     private $userRepository;
+    private $filmRepository;
 
     public function __construct()
     {
         parent::__construct();
 
         $this->userRepository = new UserRepository();
+        $this->filmRepository = new FilmRepository();
     }
 
     public function adminusers() {
@@ -23,9 +26,19 @@ class AdminUserController extends AppController {
 
     public function admindeleteuser(string $id)
     {
+        $loggedUser = unserialize($_SESSION['loggedUser']);
+        $loggedUserId = $loggedUser->getId();
+
+        if ($loggedUserId === $id) {
+            return $this->render(
+                'admin-users',
+                ['messages' => ["You can't delete your own account!"]]
+            );
+        }
+
         $this->userRepository->delete($id);
 
-        // @TODO Recalvulate averages!
+        $this->filmRepository->refreshAllAvgRate();
 
         $url = "http://$_SERVER[HTTP_HOST]";
         return header("Location: {$url}/adminusers");
